update to use httpx, refactored common code
