+++ docker-data/test-data/aveva_timeseries/test_aveva/test_aveva_20260302.csv
@@ -1,12 +1,12 @@
-time,CVOW.OSS 2 WR WAVE PERIOD,CVOW.OSS 2 WR WAVE PERIOD_IsQuestionable,CVOW.OSS 2 WR WATER LEVEL,CVOW.OSS 2 WR WATER LEVEL_IsQuestionable,CVOW.OSS 2 WR WAVE HEIGHT,CVOW.OSS 2 WR WAVE HEIGHT_IsQuestionable
-2026-03-02 17:39:15+00:00,4.940000057220459,True,-20.020000457763672,True,2.109999895095825,True
-2026-03-02 17:41:31+00:00,4.940000057220459,False,-20.020000457763672,False,2.109999895095825,False
-2026-03-02 18:35:10+00:00,4.940000057220459,True,-20.020000457763672,True,2.109999895095825,True
-2026-03-02 18:37:16+00:00,0.0,False,0.0,False,0.0,False
-2026-03-02 18:37:20+00:00,4.940000057220459,False,-20.020000457763672,False,2.109999895095825,False
-2026-03-02 19:02:16+00:00,4.940000057220459,True,-20.020000457763672,True,2.109999895095825,True
-2026-03-02 19:04:22+00:00,0.0,False,0.0,False,0.0,False
-2026-03-02 19:04:26+00:00,4.940000057220459,False,-20.020000457763672,False,2.109999895095825,False
-2026-03-02 19:47:44+00:00,4.940000057220459,True,-20.020000457763672,True,2.109999895095825,True
-2026-03-02 19:49:51+00:00,0.0,False,0.0,False,0.0,False
-2026-03-02 19:49:54+00:00,4.940000057220459,False,-20.020000457763672,False,2.109999895095825,False
+time,CVOW_OSS_2_WR_WAVE_PERIOD,CVOW_OSS_2_WR_WAVE_PERIOD_IsQuestionable,CVOW_OSS_2_WR_WATER_LEVEL,CVOW_OSS_2_WR_WATER_LEVEL_IsQuestionable,CVOW_OSS_2_WR_WAVE_HEIGHT,CVOW_OSS_2_WR_WAVE_HEIGHT_IsQuestionable
+2026-03-02T17:39:15,4.940000057220459,True,-20.020000457763672,True,2.109999895095825,True
+2026-03-02T17:41:31,4.940000057220459,False,-20.020000457763672,False,2.109999895095825,False
+2026-03-02T18:35:10,4.940000057220459,True,-20.020000457763672,True,2.109999895095825,True
+2026-03-02T18:37:16,0.0,False,0.0,False,0.0,False
+2026-03-02T18:37:20,4.940000057220459,False,-20.020000457763672,False,2.109999895095825,False
+2026-03-02T19:02:16,4.940000057220459,True,-20.020000457763672,True,2.109999895095825,True
+2026-03-02T19:04:22,0.0,False,0.0,False,0.0,False
+2026-03-02T19:04:26,4.940000057220459,False,-20.020000457763672,False,2.109999895095825,False
+2026-03-02T19:47:44,4.940000057220459,True,-20.020000457763672,True,2.109999895095825,True
+2026-03-02T19:49:51,0.0,False,0.0,False,0.0,False
+2026-03-02T19:49:54,4.940000057220459,False,-20.020000457763672,False,2.109999895095825,False
